Simplify migration mapping model creation
Simplify the code structure and remove autorelease blocks – they introduce a certain overhead and only make sense on loops with a large number of iterations and memory-intense data that should be disposed at the end of the block.

In our case, in 99% cases there will be no more than 10 iterations (attributes and relations per entity) and most results of autorelease blocks are returned and stored into the mapping model, which defies their whole purpose in the first place.